Get the JIT starting to run on macOS ARM64.
This fixes a raft of issues with hosting the JIT mapping in a
64-bit location (i.e. >4GB).
Carefully placed trampolines are used to bounce out of JIT and
into debug / interp, to avoid trying to jump to a delta that is
too great for the ARM64 architecture.
test.rom now gets a little way, and does some actual compile / run
of blocks, before hitting bus error.

+++ asm/asm_util.c
@@ -17,7 +17,7 @@ asm_make_uop0(struct asm_uop* p_uop, int32_t uopcode) {
 }
 
 void
-asm_make_uop1(struct asm_uop* p_uop, int32_t uopcode, int32_t value1) {
+asm_make_uop1(struct asm_uop* p_uop, int32_t uopcode, intptr_t value1) {
   p_uop->uopcode = uopcode;
   p_uop->value1 = value1;
   p_uop->value2 = 0;
